besspin/cyberPhys/cyberphyslib/requirements_linux.txt
transitions==0.8.7
graphviz == 0.14.2
keyboard == 0.13.5
numpy == 1.22.0
pandas == 1.1.1
psutil == 5.7.3
zmq == 0.0.0
keyboard==0.13.5
pyserial==3.5
opclib==0.0.7
python-can==3.3.4
pygame==2.0.1
setuptools_rust==0.12.1
fabric==2.6.0
